refactor: extract text insertion toast hooks
